+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support function calling.
 
-## [0.6.0] - 2024-03-XX
+## [0.6.0] - 2024-03-24
 
 ### Added